pyproject.toml
[project]
name = "blog-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.12.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
